roomcleaner-api/requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.3
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==1.0.6
Werkzeug==0.10.4
amqp==1.4.6
aniso8601==1.0.0
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
cffi==1.1.2
cryptography==0.9.3
enum34==1.0.4
gdata==2.0.18
google-api-python-client==1.4.1
gunicorn==19.3.0
httplib2==0.9.1
idna==2.0
ipaddress==1.0.7
itsdangerous==0.24
kombu==3.0.26
oauth2client==1.4.12
psycopg2==2.6.1
pyOpenSSL==0.15.1
pyasn1==0.1.8
pyasn1-modules==0.0.6
pycparser==2.14
python-dateutil==2.4.2
pytz==2015.4
redis==2.10.3
requests==2.7.0
rsa==3.1.4
simplejson==3.7.3
six==1.9.0
uritemplate==0.6
wsgiref==0.1.2
